fix RemoteNotificationService.kt and add TokenService.kt
