+++ app/build.gradle
@@ -124,6 +124,11 @@ dependencies {
     implementation 'com.github.bumptech.glide:glide:4.12.0'
     implementation 'com.google.firebase:firebase-database-ktx:20.0.0'
 
+    implementation platform('com.google.firebase:firebase-bom:31.5.0')
+
+    implementation 'com.google.firebase:firebase-messaging-ktx'
+    implementation 'com.google.firebase:firebase-analytics-ktx'
+
     // Circular ImageView
     implementation 'de.hdodenhof:circleimageview:3.1.0'
 }
