+++ app/src/main/java/com/github/ybecker/epforuml/MainActivity.kt
@@ -7,7 +7,7 @@ import androidx.appcompat.app.ActionBarDrawerToggle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.fragment.app.Fragment
-import com.github.ybecker.epforuml.Notifications.FirebaseCouldMessagingAdapter
+import com.github.ybecker.epforuml.notifications.FirebaseCouldMessagingAdapter
 import com.github.ybecker.epforuml.account.AccountFragment
 import com.github.ybecker.epforuml.account.AccountFragmentGuest
 import com.github.ybecker.epforuml.database.DatabaseManager
@@ -15,16 +15,18 @@ import com.google.android.material.navigation.NavigationView
 
 class MainActivity : AppCompatActivity() {
 
+    companion object {
+        lateinit var context: Context
+    }
+
     lateinit var toggle : ActionBarDrawerToggle
     lateinit var drawerLayout: DrawerLayout
-    lateinit var context: Context
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
         context = applicationContext
-        FirebaseCouldMessagingAdapter.init(applicationContext)
 
         // initialize DB to Mock
         //DatabaseManager.useMockDatabase()
