+++ app/src/main/java/com/github/ybecker/epforuml/SavedQuestionsFragment.kt
@@ -5,9 +5,12 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.github.ybecker.epforuml.Notifications.FirebaseCouldMessagingAdapter
+import com.github.ybecker.epforuml.notifications.FirebaseCouldMessagingAdapter
 import com.github.ybecker.epforuml.database.DatabaseManager
 import com.github.ybecker.epforuml.database.DatabaseManager.db
+import com.github.ybecker.epforuml.database.Model
+import com.github.ybecker.epforuml.notifications.RemoteNotificationService
+import com.google.firebase.messaging.RemoteMessage
 
 
 /**
@@ -22,8 +25,20 @@ class SavedQuestionsFragment : Fragment() {
     ): View? {
         // Inflate the layout for this fragment
 
-        FirebaseCouldMessagingAdapter.init(requireContext())
-        DatabaseManager.user?.let { db.addQuestion(it.userId, "-NSvYgOTEiHc6DMm994C", "newQuestion", "this is a new question", "") }
+        val message = RemoteMessage.Builder("osef")
+            .setMessageId(Integer.toString(0))
+            .addData("author", "New question from Theo")
+            .addData("title", "Ma question !")
+            .addData("text", "merci de bien vouloir repondre")
+            //.addData("icon", notification.extras.getString(NotificationCompat.EXTRA_SMALL_ICON))
+            //.addData("click_action", notification.extras.getString(NotificationCompat.EXTRA_NOTIFICATION_CLICK_ACTION))
+            .build()
+
+//        Thread(Runnable {
+//            RemoteNotificationService().onMessageReceived(message)
+//        }).start()
+
+        FirebaseCouldMessagingAdapter.sendQuestionNotifications(Model.Question("hihi", "-NT4bmu0pHU3XPM6qx7W","0", "NEW QUESTION", "JSPR QUE CA MARCHE","", emptyList()))
 
         return inflater.inflate(R.layout.fragment_saved_questions, container, false)
