+++ app/src/main/java/com/github/ybecker/epforuml/notifications/RemoteNotificationService.kt
@@ -1,16 +1,15 @@
 
-package com.github.ybecker.epforuml.Notifications
+package com.github.ybecker.epforuml.notifications
 
 import android.Manifest
 import android.app.NotificationChannel
 import android.app.NotificationManager
-import android.content.ContentValues.TAG
 import android.content.pm.PackageManager
 import android.os.Build
-import android.util.Log
 import androidx.core.app.ActivityCompat
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
+import com.github.ybecker.epforuml.MainActivity
 import com.github.ybecker.epforuml.R
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
@@ -21,49 +20,37 @@ class RemoteNotificationService : FirebaseMessagingService() {
     override fun onMessageReceived(remoteMessage: RemoteMessage) {
         super.onMessageReceived(remoteMessage)
 
-        // Récupérer le titre et le corps du message
+        // Get data from the message
         val author = remoteMessage.data["author"]
-        val title = remoteMessage.notification?.title
-        val body = remoteMessage.notification?.body
+        val title = remoteMessage.data["title"]
+        val text = remoteMessage.data["text"]
         val channelId = "my_channel_id"
 
-        // Créer le canal de notification (disponible uniquement pour Android Oreo et versions supérieures)
+        // creat new channel
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             val channel = NotificationChannel(channelId, "My Channel", NotificationManager.IMPORTANCE_DEFAULT)
             val notificationManager: NotificationManager =
-                FirebaseCouldMessagingAdapter.context.getSystemService(NOTIFICATION_SERVICE) as NotificationManager
-            notificationManager?.createNotificationChannel(channel)
+                MainActivity.context.getSystemService(NOTIFICATION_SERVICE) as NotificationManager
+            notificationManager.createNotificationChannel(channel)
         }
 
         // Afficher une notification
-        val notification = NotificationCompat.Builder(FirebaseCouldMessagingAdapter.context, channelId)
+        val notification = NotificationCompat.Builder( MainActivity.context, channelId)
         .setSmallIcon(R.drawable.nav_chat)
         .setContentTitle(author)
         .setContentText(title)
         //.setLargeIcon(bitmap)
-        .setStyle(NotificationCompat.BigTextStyle().bigText(title+"\n"+body))
+        .setStyle(NotificationCompat.BigTextStyle().bigText(title+"\n"+text))
         .setPriority(NotificationCompat.PRIORITY_DEFAULT)
         .build()
 
-        val notificationManager = NotificationManagerCompat.from(this)
+        val notificationManager = NotificationManagerCompat.from(MainActivity.context)
         if (ActivityCompat.checkSelfPermission(
-                this,
+                MainActivity.context,
                 Manifest.permission.POST_NOTIFICATIONS
             ) == PackageManager.PERMISSION_GRANTED
         ) {
             notificationManager.notify(0, notification)
         }
-    }
-
-    override fun onNewToken(token: String) {
-        super.onNewToken(token)
-        Log.d(TAG, "Refreshed token: $token")
-
-        //TODO
-        // Enregistrement du token dans la base de données
-        // ...
-
-        // Envoi du token au serveur d'application si nécessaire
-        // ...
     }
 }
